implement constrained restore that only walks paths specified in the --include pattern
If the user is doing restore using "--include" paths, and all of the
given paths are absolute (begin with '/'), and no glob match chars are
present, we can prevent walking the entire restore tree, and instead
visit only subdirs that match the given "--include" paths.

This is a huge speedup for large/remote repos. In my testing with s3,
I went 53m18s to 1m17s. Prior to this patch, the restore often failed
to complete at all due to timeouts.