Skip subdirs on local backends for keys
In cases where there are subdirectories present in the keys/ dir the
path returned is stripped off the component.

Keys should reside only in the top level of keys/ so skip the
subdirectories for now.

The use case causing this patch is a local storage mounted via cifs on
some webdav server that litters .DAV subdirectories everywhere.