cmd: restic: clean up sort.SliceStable usage
The older API is still available in the Go stdlib (and will remain there
forever, thanks to Go's backwards compatibility guarantees). By
switching back to the pre-1.8 API this allows restic to be built on
older distributions that don't have a newer Go compiler (though it turns
out some of our vendors use the newer API, this is something
distributions can patch out themselves).

It also (in my opinion) makes the code much more readable (since
time.Time implements sort.Interface, there's no need to explicitly
implement it using .After and .Before). We've been carrying this patch
in openSUSE for a couple of releases and nobody seems to have noticed an
issue with it.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>