Disable scanner for quiet backup
The scanner's only purpose is to provide totals for file and directory
counts and size. These are used to provide a sensible progress bar for
the user. When running restic in quiet mode, no output is produced and
thus there is no need to run the scanner. This saves the processing and
syscalls required to walk over potentially millions of files.