repository: Reduce memory usage of index entries
Optimize for the default case where each blob is stored in exactly
one blob. Duplicate blobs can only appear if a backup manages to upload
a pack file but no corresponding index. In addition, the prune command
will remove these duplicated blobs. Therefore, only a small fraction of
blobs will ever be present in multiple packs.

Previously the blob-handle-to-pack-id-map stored a slice pointing to a
list of pack entries. The slice part which allows storing multiple
entries is, however, rarely needed. This patch replaces it with a
pointer to a single pack entry descriptor and a separate map to handle
"overflows", that is multiple pack entries for a single blob handle.

The simpler option of directly storing the entry descriptor in the map
would require more memory in the common case: A map contains a lot of
unused buckets which also consume memory. On amd64 a pointer to an
entry is 8 bytes, a slice is 24 bytes in size, and a pack entry requires
48 bytes. Maps in Go have a max. load factor L of 6.5/8 which after
growing will result in a load factor L of 3.25/8. With N being the
number of map slots this yields the following memory usage:

Previous: N*24 + N*L*48
Pointer: N*8 + N*L*48
Inline: N*48

The Pointer variant obviously needs less memory than Previous.
Inline needs more memory than the Pointer variant if
Inline - Pointer > 0 <=> N * (48 - 8 - L*48) > 0 <=> L < 5/6
As 5/6 > max. load factor 6.5/8 the Pointer variant always requires less
memory.

Note that shrinking a pack entry to 40 bytes would make the Inline
variant preferable for L > 2/3. The Pointer variant still needs to
allocate 48 bytes as the Go memory allocator has no pool for 40byte
objects and therefore the next biggest, that is 48 bytes, will be used.
L > 2/3 would cover approximately one third of the usual load factor
range, such that the Pointer variant would still be preferable on
average.