Add new flag masterkeyfile
Add the new flag `--masterkeyfile` which allows to directly use the
master key from a file.
When `--masterkeyfile` is used from `restic init` the master key is
written to the given file and no key is created in the repo.
For all other commands, `--masterkeyfile` uses the master key in the
given file. No password is needed.