repository: Increase max blobs for preliminary index to 3000
Each index file is currently loaded into a separate hashmap. Go hashmaps
use a load factor of 6.5/8 and allocate an array whose size is a power
of two. For 2000 entries this requires a array with 4096 slots which
wastes a lot of memory. The maximum number of entries for that array
size would be 3328. However, leave some headroom as the index_uploader
only runs every 30 seconds.