Reworked print functions to better match verbosity levels
Added new functions PrintDef and Verboseff to print output when the verbosity
level is respectively greater than or equal to 1 (default) or greater than or
equal to 3 ("--verbose=2"), and changed Verbosef to print output when verbosity
is greater than or equal to 2 ("--verbose").

Converted all calls to Verbosef to one of the new functions and removed some direct
checks of the verbosity level.