Fix runtime.KeepAlive usage in preallocateFile
When fst is a struct, what runtime.KeepAlive receives is a copy pointed
to by an interface{}. If I've got the unsafe.Pointer rules right, the
compiler is allowed to decide that neither FcntlInt call can see the
struct because all it they get is an int. It's then allowed to reorder
the code to first copy the struct, then drop the memory for the
original, then call FcntlInt.

In any case, this matches the example in the runtime package more
closely.