Add --device-map flag to restic backup
When backing up from a snapshot (ZFS or btrfs as an example), each
snapshot will be mounted by the filesystem as a different DeviceID

This causes restic to upload the directory structure for each new
snapshot, even though the structure is identical. This commit adds a new
flag to prevent this behavior and allow restic to re-use the previously
uploaded directory structure.

The concept is simple. When calling restic backup, pass the --device-map
src:dest to change any node with a DeviceID of src to behave as if it
were dest

Luckily, nodeFromFileInfo is the only place the DeviceID is read in the
backup codepath, so it's straightforward to shim the DeviceID here

Closes #3041