Add option to follow symbolic links
A new flag --follow-symlinks was added to the backup command (default off).  When enabled, restic will dereference any symbolic link encountered to its destination file or folder instead.  This allows the destination of symbolic links to be backed up as if they were standard files or folders.  In order to prevent infinite loops or backup of unwanted symlinks, use the --exclude options.