fuse: Concurrent readahead
Add concurrent readahead to fuse.openFile.Read. Every read call now
spans a goroutine (up to some memory limit) that reads the first block
after the section it is told to read, in the expectation that the next
call will need that block.

I ran two small benchmarks on a compressed repository, using a local
on-disk cache.

Benchmark one: sha256sum on a 753MiB file. Best of three runs, after a
single run to warm up the caches.

    # 59a90943bb0c780fa69fafe752b2be4a98014397
    real    0m17.266s
    user    0m5.255s
    sys     0m0.391s

    # This PR
    real    0m13.808s
    user    0m4.662s
    sys     0m0.298s

That's -20%, or a 25% speed increase.

Benchmark two: sha256sum of the same file in parallel with a 703MiB
file, with which it shares no blocks. This simulates the habit of some
file managers of copying two files concurrently when copying a
directory.

    # 59a90943bb0c780fa69fafe752b2be4a98014397
    real    0m15.789s
    user    0m8.846s
    sys     0m0.676s

    # This PR
    real    0m13.643s
    user    0m8.776s
    sys     0m0.606s

That's -14%, or a 16% speed increase.