dry-run: introduce environment RESTIC_DRY_RUN
Various commands support the --dry-run flag. When restic is used as part
of a scripted chain, adding/removing "--dry-run" to check the chain is
tedious.

This commit introduces the evaluation of the environment variable
RESTIC_DRY_RUN. If set, "--dry-run=true" is assumed. It applies
to commands "backup", "forget", "prune" and "rewrite".