JSON logging: output "unchanged" only if level -vv
This patch changes json logging so that it outputs
"action": "unchanged" only, if the log level is at least -vv.

Other actions like "new" and "modified" will still be printed
on log-level -v. That way it is possible to opt out of the
large amount of messages regarding "unchanged" files and dirs,
which might not be of interest and still get all the information
regarding what has changed to the previous snapshot.